+++ secrets-engines-database.go
@@ -1,9 +1,7 @@
 package main
 
 import (
-	// "fmt"
-	// "reflect"
-	log "github.com/Sirupsen/logrus"
+	log "github.com/sirupsen/logrus"
 	"io/ioutil"
 	"path/filepath"
 )